services/plans-service/pyproject.toml
[project]
name = "plans-service"
version = "0.1.0"
description = "Calendar plans microservice"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "pydantic>=2.7",
  "requests>=2.31",
  "python-dotenv>=1.0",
  "psycopg2-binary>=2.9",
  "pydantic-settings>=2.2",
  "httpx>=0.27",  # Added for making HTTP requests
  "tenacity>=8.3",  # Added for retry functionality
  "asyncpg>=0.30",  # Added for async PostgreSQL support
  "structlog>=24.0.0,<25.0.0",
  "asgi-correlation-id>=4.3.1,<5.0.0",
  "prometheus-fastapi-instrumentator>=6.1.0,<7.0.0",
  "prometheus-client>=0.16.0,<1.0.0",
  "sentry-sdk[fastapi]>=2.0.0,<3.0.0",
  "redis>=5",
  "celery[redis]>=5.4.0,<6.0.0",
  "backend-common @ file:///libs/backend-common",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "httpx>=0.27",
  "pytest-asyncio>=0.23",
]
dev = [
  "ruff>=0.5",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["plans_service"]
package-dir = { "" = "." }
